refactor: consolidate message models into chat_manager and prompt_generator, remove unused message_model
