+++ custom_components/openai_conversation_for_rs/chat_manager.py
@@ -1,7 +1,25 @@
 """Chat manager module."""
 
-from message_model import BaseMessage
+from typing import Optional
+
 from prompt_manager import ClientCache
+from pydantic import BaseModel
+
+
+class BaseMessage(BaseModel):
+    """Base message model"""
+
+    id: Optional[int] = None
+    role: str
+    content: str
+
+    def to_dict(self) -> dict:
+        """Convert the message to a dictionary"""
+        return {
+            "id": self.id,
+            "role": self.role,
+            "content": self.content,
+        }
 
 
 class ChatCache(ClientCache):
